Generate META.json
Rather than modify it in the repository and potentially create weird
conflicts. Generate it from an input file, instead.

Use `git config --add safe.directory` to tell Git to trust the current
directory, even though it is used by a different user. This is required
to work around pgxn/docker-pgxn-tools#5.

But also use the newly-released `pgxn/pgxn-tools:1.4` image to build the
distribution, so as to use the new `GIT_BUNDLE_OPTS` variable to add the
generated `META.json` file to the bundle. It would otherwise be excluded
because `git archive` includes only committed files.